+++ src/main/native/glue/s/SoftBodyMotionProperties.cpp
@@ -195,7 +195,7 @@ JNIEXPORT jint JNICALL Java_com_github_stephengold_joltjni_SoftBodyMotionPropert
     const Vec3 comLocation(x, y, z);
     const Array<SoftBodyVertex>& vertices = pProperties->GetVertices();
     const Array<SoftBodyVertex>::size_type numVertices = vertices.size();
-    for (int i = 0; i < numVertices && bufferPosition + 2 < capacityFloats; ++i) {
+    for (size_t i = 0; i < numVertices && bufferPosition + 2 < capacityFloats; ++i) {
         const SoftBodyVertex &vertex = vertices[i];
         if (vertex.mInvMass == 0) {
             const Vec3& vertexLocation = vertex.mPosition + comLocation;
@@ -225,7 +225,7 @@ JNIEXPORT jint JNICALL Java_com_github_stephengold_joltjni_SoftBodyMotionPropert
     const Vec3 comLocation(x, y, z);
     const Array<SoftBodyVertex>& vertices = pProperties->GetVertices();
     const Array<SoftBodyVertex>::size_type numVertices = vertices.size();
-    for (int i = 0; i < numVertices && bufferPosition + 2 < capacityFloats; ++i) {
+    for (size_t i = 0; i < numVertices && bufferPosition + 2 < capacityFloats; ++i) {
         const SoftBodyVertex &vertex = vertices[i];
         const Vec3& vertexLocation = vertex.mPosition + comLocation;
         pStoreFloats[bufferPosition++] = vertexLocation.GetX();
